Adding code for cleaning CA historical catch data
Preliminary analysis of CA historical catch data and comparison to 2017 assessment historical data. Also includes interpolation of data from 1888 to 1915, which was suggested by the STAR panel from the 2017 update. Overall, the CA historical data matches well with the data used in the 2017 update, with the exception of a few years, which I still need to figure out why this is happening.