Adding code for cleaning CA historical catch data (#31)
* Adding code for cleaning CA historical catch data

Preliminary analysis of CA historical catch data and comparison to 2017 assessment historical data. Also includes interpolation of data from 1888 to 1915, which was suggested by the STAR panel from the 2017 update. Overall, the CA historical data matches well with the data used in the 2017 update, with the exception of a few years, which I still need to figure out why this is happening.

* Providing yelloweye catches from 2017 assessment

Adding the CSV of the catches used in the 2017 assessment that Vlada sent over to double check with. I also switched the start year to 1889.

* Readding catch data from 2017 assessment

Re-adding the catch data from 2017 assessment into the raw folder that Vlada sent to compare historical catches to.

* Switching 2017 catches to processed folder

Switching the 2017 catches to processed folder. This is a csv that Vlada sent to use to check the catch data.