Fishery dep indices (#125)
* Create AtSea-CPUE_filtered_457_2024.csv

Finished the filtered drift-level ORFS dataset - ready to be put into sdmTMB

* Delete AtSea-CPUE_filtered_457_2024.csv

moved ORFS dataset and added ORBS (both are aggregated to the trip-level and have preliminary filters applied)

* Rscripts for ORBS/ORFS

adding the scripts to process the ORFS and ORBS datasets in sdmTMB

* removing outdated R scripts - will replace

* updated ORBS index script and helper functions

* Updated code and added the (likely) final index for ORBS

* provide ORFS index

such as it is...

* Added ORBS index tables and figures

Added a standard set of figures and tables for the updated ORBS index.

* Adding files for ORFS index

Because I should have done it earlier... tables and figures for the ORFS (onboard observer) index.

---------

Co-authored-by: Claire Rosemond <claire.rosemond@noaa.gov>