Updated sensitivities with new base!  WOOHOO
and its really late, but all the sensitivities are there and i might need to do another plot or table for it tomorrow but i'm going to bed now!  :)  want to know how many times i had to re-clone? i will buy you both that many beers when we're done